feat: persist transactions in the database
createConnection has been renamed to createDefaultConnection and it always returns a connection named "default". TypeORM looks up for a connection named "default" when not specified. If we create a connection named "test-connection" it'll not be found by TypeORM when repositories are used.
Added a temporary connection for the test env in ormconfig.json